LayoutTests/ietestcenter/Javascript/15.2.3.9-1-expected.txt

15.2.3.9-1 (TestCases/chapter15/15.2/15.2.3/15.2.3.9/15.2.3.9-1.js) Object.freeze throws TypeError if type of first param is not Object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL ES5Harness.preconditionPassed should be true (of type boolean). Was undefined (of type undefined).
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
